add test

+++ src/main/java/com/hxj/app/pack/ModulePackager.java
@@ -1,16 +1,17 @@
-package com.hxj.app.APICloudModulePackager;
+package com.hxj.app.pack;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import com.google.gson.stream.JsonReader;
+import com.hxj.app.pack.ex.PackException;
 import com.hxj.common.file.FileUtils;
 
 /**
@@ -20,47 +21,59 @@ import com.hxj.common.file.FileUtils;
  */
 public class ModulePackager {
 
+	private final Logger	LOG	= LoggerFactory.getLogger(ModulePackager.class);
+
 	/**
 	 * 项目路径
 	 */
-	private String		projectPath;
+	private String			projectPath;
 	/**
 	 * 模块名称
 	 */
-	private String		moduleName;
+	private String			moduleName;
 	/**
 	 * 打包输出路径
 	 */
-	private String		outputPath;
+	private String			outputPath;
 	/**
 	 * 源码列表
 	 */
-	private List<File>	sourceList;
+	private List<File>		sourceList;
 	/**
 	 * 引用库列表
 	 */
-	private List<File>	libList;
+	private List<File>		libList;
 	/**
 	 * 资源列表
 	 */
-	private List<File>	resList;
+	private List<File>		resList;
 	/**
 	 * module.json
 	 */
-	private File		moduleFile;
+	private File			moduleFile;
+	/**
+	 * AndroidManifest.xml
+	 */
+	private File			manifestFile;
 
 	/**
 	 * 执行模块打包
+	 * @throws PackException 
 	 */
-	public void doPack() {
+	public void doPack() throws PackException {
+		LOG.info("do pack");
+		checkPackInfo();
 		try {
 			// 1.创建输出目录结构
 			File outFile = new File(outputPath);
-			if (outFile.exists()) {
+			if (outFile.exists() && outFile.isDirectory()) {
+				LOG.debug("out dir exists, delete inner files.");
 				FileUtils.deleteDirFiles(outFile);
 			} else {
-				if (!outFile.createNewFile()) {
+				LOG.debug("create out dir");
+				if (!outFile.mkdirs()) {
 					// create fail
+					LOG.error("out dir create fail");
 					return;
 				}
 			}
@@ -78,9 +91,11 @@ public class ModulePackager {
 			moduleFile.createNewFile();
 
 			// 2.复制文件
+			// 2.3 res
+			FileUtils.copyFile(this.manifestFile, new File(resFile,this.manifestFile.getName()));
 			// 2.4 module.json
 			FileUtils.copyFile(this.moduleFile, moduleFile);
-			String str = FileUtils.readJSONFile(this.moduleFile);
+			String str = FileUtils.readFileToString(this.moduleFile);
 			JsonElement json = new JsonParser().parse(str);
 			JsonElement modules = json.getAsJsonObject().get("modules");
 			if (modules != null) {
@@ -103,6 +118,23 @@ public class ModulePackager {
 			e.printStackTrace();
 		}
 
+	}
+	
+	/**
+	 * 检查打包信息
+	 * @return
+	 * @throws PackException 
+	 */
+	private void checkPackInfo() throws PackException {
+		if (!FileUtils.isDirExists(projectPath)) {
+			throw new PackException("error project path:" + projectPath);
+		}
+		if (!FileUtils.isFileExists(moduleFile)) {
+			throw new PackException("can not file module.json:" + moduleFile);
+		}
+		if (!FileUtils.isFileExists(manifestFile)) {
+			throw new PackException("can not file AndroidManifest.xml:" + manifestFile);
+		}
 	}
 
 	public String getProjectPath() {
@@ -152,4 +184,24 @@ public class ModulePackager {
 	public void setResList(List<File> resList) {
 		this.resList = resList;
 	}
+
+	public File getModuleFile() {
+		return moduleFile;
+	}
+
+	public void setModuleFile(File moduleFile) {
+		this.moduleFile = moduleFile;
+	}
+
+	public File getManifestFile() {
+		return manifestFile;
+	}
+
+	public void setManifestFile(File manifestFile) {
+		this.manifestFile = manifestFile;
+	}
+	
+	
+	
+	
 }
